Extract document nodes package
