+++ src/main/java/com/code4ro/legalconsultation/converters/DocumentConsolidatedMapper.java
@@ -1,5 +1,6 @@
 package com.code4ro.legalconsultation.converters;
 
+import com.code4ro.legalconsultation.document.node.mapper.DocumentNodeMapper;
 import com.code4ro.legalconsultation.model.dto.DocumentConsolidatedDto;
 import com.code4ro.legalconsultation.model.persistence.DocumentConsolidated;
 import com.code4ro.legalconsultation.user.mapper.UserMapper;
