+++ src/main/java/com/code4ro/legalconsultation/document/node/mapper/DocumentNodeMapper.java
@@ -1,9 +1,9 @@
-package com.code4ro.legalconsultation.converters;
+package com.code4ro.legalconsultation.document.node.mapper;
 
-import com.code4ro.legalconsultation.model.dto.documentnode.DocumentNodeCreateDto;
-import com.code4ro.legalconsultation.model.dto.documentnode.DocumentNodeDto;
-import com.code4ro.legalconsultation.model.dto.documentnode.DocumentNodeSimpleDto;
-import com.code4ro.legalconsultation.model.persistence.DocumentNode;
+import com.code4ro.legalconsultation.document.node.model.dto.DocumentNodeCreateDto;
+import com.code4ro.legalconsultation.document.node.model.dto.DocumentNodeDto;
+import com.code4ro.legalconsultation.document.node.model.dto.DocumentNodeSimpleDto;
+import com.code4ro.legalconsultation.document.node.model.persistence.DocumentNode;
 import org.mapstruct.Mapper;
 
 @Mapper(componentModel = "spring", uses = {NumberOfCommentsMapper.class})
