+++ src/main/java/com/code4ro/legalconsultation/document/node/model/dto/DocumentNodeCreateDto.java
@@ -1,14 +1,13 @@
-package com.code4ro.legalconsultation.model.dto.documentnode;
+package com.code4ro.legalconsultation.document.node.model.dto;
 
-import com.code4ro.legalconsultation.model.dto.BaseEntityDto;
-import com.code4ro.legalconsultation.model.persistence.DocumentNodeType;
+import com.code4ro.legalconsultation.document.node.model.persistence.DocumentNodeType;
 import lombok.Data;
 
 import javax.validation.constraints.Size;
 import java.util.UUID;
 
 @Data
-public class DocumentNodeSimpleDto extends BaseEntityDto {
+public class DocumentNodeCreateDto {
     private UUID parentId;
     private DocumentNodeType documentNodeType;
     @Size(max = 255)
