+++ src/main/java/com/code4ro/legalconsultation/document/node/model/dto/DocumentNodeSimpleDto.java
@@ -1,13 +1,14 @@
-package com.code4ro.legalconsultation.model.dto.documentnode;
+package com.code4ro.legalconsultation.document.node.model.dto;
 
-import com.code4ro.legalconsultation.model.persistence.DocumentNodeType;
+import com.code4ro.legalconsultation.model.dto.BaseEntityDto;
+import com.code4ro.legalconsultation.document.node.model.persistence.DocumentNodeType;
 import lombok.Data;
 
 import javax.validation.constraints.Size;
 import java.util.UUID;
 
 @Data
-public class DocumentNodeCreateDto {
+public class DocumentNodeSimpleDto extends BaseEntityDto {
     private UUID parentId;
     private DocumentNodeType documentNodeType;
     @Size(max = 255)
