+++ src/main/java/com/code4ro/legalconsultation/document/node/model/persistence/DocumentNode.java
@@ -1,5 +1,6 @@
-package com.code4ro.legalconsultation.model.persistence;
+package com.code4ro.legalconsultation.document.node.model.persistence;
 
+import com.code4ro.legalconsultation.model.persistence.BaseEntity;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.ToString;
