+++ src/main/java/com/code4ro/legalconsultation/document/node/service/DocumentNodeService.java
@@ -1,7 +1,7 @@
-package com.code4ro.legalconsultation.service.api;
+package com.code4ro.legalconsultation.document.node.service;
 
-import com.code4ro.legalconsultation.model.dto.documentnode.DocumentNodeSimpleDto;
-import com.code4ro.legalconsultation.model.persistence.DocumentNode;
+import com.code4ro.legalconsultation.document.node.model.dto.DocumentNodeSimpleDto;
+import com.code4ro.legalconsultation.document.node.model.persistence.DocumentNode;
 
 import java.util.UUID;
