+++ src/main/java/com/code4ro/legalconsultation/service/impl/pdf/parser/DocumentParser.java
@@ -1,7 +1,7 @@
 package com.code4ro.legalconsultation.service.impl.pdf.parser;
 
-import com.code4ro.legalconsultation.model.persistence.DocumentNode;
-import com.code4ro.legalconsultation.model.persistence.DocumentNodeType;
+import com.code4ro.legalconsultation.document.node.model.persistence.DocumentNode;
+import com.code4ro.legalconsultation.document.node.model.persistence.DocumentNodeType;
 import org.springframework.stereotype.Component;
 
 import java.util.regex.Matcher;
