+++ src/test/java/com/code4ro/legalconsultation/comment/controller/CommentControllerIntegrationTest.java
@@ -7,9 +7,9 @@ import com.code4ro.legalconsultation.comment.service.CommentService;
 import com.code4ro.legalconsultation.common.controller.AbstractControllerIntegrationTest;
 import com.code4ro.legalconsultation.config.security.CurrentUserService;
 import com.code4ro.legalconsultation.comment.factory.CommentFactory;
-import com.code4ro.legalconsultation.factory.DocumentNodeFactory;
+import com.code4ro.legalconsultation.document.node.factory.DocumentNodeFactory;
 import com.code4ro.legalconsultation.model.persistence.ApplicationUser;
-import com.code4ro.legalconsultation.model.persistence.DocumentNode;
+import com.code4ro.legalconsultation.document.node.model.persistence.DocumentNode;
 import com.code4ro.legalconsultation.comment.repository.CommentRepository;
 import org.junit.Before;
 import org.junit.Test;
