+++ src/test/java/com/code4ro/legalconsultation/document/node/controller/DocumentNodeControllerTest.java
@@ -1,10 +1,11 @@
-package com.code4ro.legalconsultation.controller;
+package com.code4ro.legalconsultation.document.node.controller;
 
-import com.code4ro.legalconsultation.converters.DocumentNodeMapper;
-import com.code4ro.legalconsultation.converters.DocumentNodeMapperImpl;
-import com.code4ro.legalconsultation.model.dto.documentnode.DocumentNodeCreateDto;
-import com.code4ro.legalconsultation.model.persistence.DocumentNode;
-import com.code4ro.legalconsultation.service.api.DocumentNodeService;
+import com.code4ro.legalconsultation.document.node.controller.DocumentNodeController;
+import com.code4ro.legalconsultation.document.node.mapper.DocumentNodeMapper;
+import com.code4ro.legalconsultation.document.node.mapper.DocumentNodeMapperImpl;
+import com.code4ro.legalconsultation.document.node.model.dto.DocumentNodeCreateDto;
+import com.code4ro.legalconsultation.document.node.model.persistence.DocumentNode;
+import com.code4ro.legalconsultation.document.node.service.DocumentNodeService;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
